feat&refactor ModalView 생성 후 상속 구조로 변경 및 파일 구조 정리
- autoCompleteView와 recentSearchView에서 중복되는 부분을 ModalView로 만들고 상속 및 오버라이딩 받도록 수정
- index.html에서 하나의 script파일만 로드하도록 파일 구조 정리
