+++ UI_component/carouselUI/controller.js
@@ -1,7 +1,7 @@
 "use strict";
 import Carousel from "./carousel.js";
 import Pagination from "./pagination.js";
-import $ from "../../mylibrary.js";
+import $ from "../../Utills/mylibrary.js";
 
 class Controller {
   constructor(obj) {
