+++ UI_component/carouselUI/pagination.js
@@ -1,5 +1,5 @@
 "use strict";
-import $ from "../../mylibrary.js";
+import $ from "../../Utills/mylibrary.js";
 
 class Pagination {
   constructor(inputTag) {
