+++ UI_component/searchUI/app.js
@@ -3,16 +3,18 @@ import SearchView from "./searchView.js";
 import AutoCompleteView from "./autoCompleteView.js";
 import RecentSearchView from "./recentSearchView.js";
 
-const searchView = new SearchView([".search-form", ".btn-search"]);
-const autoCompleteView = new AutoCompleteView(".modal-autocomplete");
-const recentSearchView = new RecentSearchView(".modal-recent");
+export default function loadSearchUI() {
+  const searchView = new SearchView([".search-form", ".btn-search"]);
+  const autoCompleteView = new AutoCompleteView(".modal-autocomplete");
+  const recentSearchView = new RecentSearchView(".modal-recent");
 
-const controller = new Controller({
-  searchView,
-  autoCompleteView,
-  recentSearchView
-});
+  const controller = new Controller({
+    searchView,
+    autoCompleteView,
+    recentSearchView
+  });
 
-window.addEventListener("DOMContentLoaded", () => {
-  controller.getData("../../resources/data/searchData.json");
-});
+  window.addEventListener("DOMContentLoaded", () => {
+    controller.getData("../../resources/data/searchData.json");
+  });
+}
