+++ app.js
@@ -1,0 +1,5 @@
+import loadCarouselUI from "./UI_component/carouselUI/app.js";
+import loadSearchUI from "./UI_component/searchUI/app.js";
+
+loadCarouselUI();
+loadSearchUI();
