+++ index.html
@@ -57,8 +57,6 @@
         </div>
       </div>
     </div>
-    <!-- <script type="module" src="./app.js"></script> -->
-    <script type="module" src="./UI_component/carouselUI/app.js"></script>
-    <script type="module" src="./UI_component/searchUI/app.js"></script>
+    <script type="module" src="./app.js"></script>
   </body>
 </html>
